moving calculate-date.ts to ./common/src/utils/ folder
