+++ mobile/src/utils/functions/index.ts
@@ -1,2 +1,2 @@
 // export utils functions
-export { calculateDate } from "./calculate-date";
+export { calculateDate } from "../../.common/utils/calculate-date";
